Since this project has an LLVM dependency, it's being moved to being Clang compatible, meaning some significant work needs to be done to clean up the iterators, since LLVM has a higher required iterator debug level to run. Maybe not but I have spent too many hours trying to set LLVM up on Windows and I just wanna work on this.
Currently, the lexing is fixed, AST generation will be fixed and finished at some point in the future. Lexing of course is not perfect but it can parse simple code which is enough for now (see the short demo in main.cpp, use debugger to view lex_token vector).

Also also the CMakeLists.txt is now provided, I did not initially think it was standard to include it in a repo, but I was wrong.