Codegen is now done via polymorphic node types for future optimized node types to be more easily slotted into the codegen process.
God, I do not want to talk about how much of a pain in the ass this was to implement. I spent a good two hours staring at linking errors because I forward declared a struct as a class, which CLion did not seem to have a problem with for some reason.